+++ src/backend/commands/admin-command.manager.ts
@@ -21,6 +21,7 @@ import {GeoManagerCommand} from './geo-manager.command';
 import {PageManagerCommand} from '@dps/mycms-server-commons/dist/pdoc-backend-commons/commands/pdoc-manager.command';
 import {PDocLoaderCommand} from '@dps/mycms-server-commons/dist/pdoc-backend-commons/commands/pdoc-loader.command';
 import {PDocConverterCommand} from '@dps/mycms-server-commons/dist/pdoc-backend-commons/commands/pdoc-converter.command';
+import {PdfManagerCommand} from './pdf-manager.command';
 
 // tslint:disable-next-line:no-empty-interface
 export interface AdminCommandConfigType extends CommonAdminCommandConfigType {
@@ -43,6 +44,7 @@ export class AdminCommandManager extends CommonAdminCommandManager<AdminCommandC
             'objectDetectionManager': new ObjectDetectionManagerCommand(),
             'geoManagerCommand': new GeoManagerCommand(),
             'pageManager': new PageManagerCommand(),
+            'pdfManager': new PdfManagerCommand(),
             'facetCacheManager': new FacetCacheManagerCommand(),
             'dbAdmin': new DbAdminCommand(),
             'dbMigrate': new DbMigrateCommand(),